Only report first result from integration tests
Once again casper has proven it's shittyness as a test runner.
The entire file is run unless an exception is raised. Failing an
assertion does not raise an exception. Yes, I know. This makes
perfect sense. So your tests will dilgently continue to fail.
I cannot think of a fix for this at the moment. This commit
is a hack to only report the first result from a single test.
I will work on a better solution.