Integration tests execute inside the browser
This commit drastically changes the way integration tests are written.
Previously they were casper scripts that interacted with a remote
browser. This created two worlds: the one where tests ran and where the
code was actually running. This is hard to grasp and hard for beginners.
This also means you must evaluate code in the remote process all the
time to get things to assert in your tests. This makes tests nasty.
Also, it binds your integration tests (the most important ones) to
CasperJS. This is unacceptable. This commit changes this.

Integration tests are now written in qunit. The test code is injected
into a booted app and the test is started. You can ineract directly
with your app code since now they are in the same process. You can use
jQuery to ineract with the page and trigger events. This approach is
cleaner and not bound to any specific test runner. Casper is still used
to interact and inject test code. The browser is refreshed before each
test run. This ensures that each test starts with a completely clean
state and saves developers from potentially costly mistakes.

Casper scripts are still support. Place them in `test/casper`.

This is for the first step to getting tests that can run in real
browsers.