Generate index.html.erb with new apps
Previously index.html was generated inside a rack middleware when
serving the app. This created a problem: there would be no index.html
when running iridium compile for stock apps. This means they cannot be
deployed without iridium. This change includes the same index.html ERB
template with generated app. This also takes some of the magic away
which may be a good thing.