Depend on handlebars-source
Barber has a hard dependency on handlebars-source. At this point it
depends on an RC version. This version is specified in the gemspec so
iridium will install correctly and precompiling will work.

The files in /vendor are replaced in favor of using what's bundled with
handlebars-source. Users can still replace handlebars and
handlebars.runtime in their vendor/javascripts folder to upgrade.
handlebars-source is only used when generating an app.